move impl FromRequest for Ruma up one level; some cleanup
